+++ src/server/table_stats.h
@@ -1,0 +1,142 @@
+// Copyright (c) 2017, Xiaomi, Inc.  All rights reserved.
+// This source code is licensed under the Apache License Version 2.0, which
+// can be found in the LICENSE file in the root directory of this source tree.
+
+#pragma once
+
+struct table_stats
+{
+    table_stats(const std::string &app_name) { this->app_name = app_name; }
+
+    double get_total_read_qps() const
+    {
+        return total_get_qps + total_multi_get_qps + total_scan_qps;
+    }
+
+    double get_total_write_qps() const
+    {
+        return total_put_qps + total_multi_put_qps + total_remove_qps + total_multi_remove_qps +
+               total_incr_qps + total_check_and_set_qps + total_check_and_mutate_qps;
+    }
+
+    void aggregate(const row_data &row)
+    {
+        total_get_qps += row.get_qps;
+        total_multi_get_qps += row.multi_get_qps;
+        total_put_qps += row.put_qps;
+        total_multi_put_qps += row.multi_put_qps;
+        total_remove_qps += row.remove_qps;
+        total_multi_remove_qps += row.multi_remove_qps;
+        total_incr_qps += row.incr_qps;
+        total_check_and_set_qps += row.check_and_set_qps;
+        total_check_and_mutate_qps += row.check_and_mutate_qps;
+        total_scan_qps += row.scan_qps;
+        total_recent_read_cu += row.recent_read_cu;
+        total_recent_write_cu += row.recent_write_cu;
+        total_recent_expire_count += row.recent_expire_count;
+        total_recent_filter_count += row.recent_filter_count;
+        total_recent_abnormal_count += row.recent_abnormal_count;
+        total_recent_write_throttling_delay_count += row.recent_write_throttling_delay_count;
+        total_recent_write_throttling_reject_count += row.recent_write_throttling_reject_count;
+        total_storage_mb += row.storage_mb;
+        total_storage_count += row.storage_count;
+        total_rdb_block_cache_hit_count += row.rdb_block_cache_hit_count;
+        total_rdb_block_cache_total_count += row.rdb_block_cache_total_count;
+        total_rdb_index_and_filter_blocks_mem_usage += row.rdb_index_and_filter_blocks_mem_usage;
+        total_rdb_memtable_mem_usage += row.rdb_memtable_mem_usage;
+
+        // get max_total_qps、min_total_qps and the id of this partition which has max_total_qps
+        double row_total_qps = row.get_total_qps();
+        min_total_qps = std::min(min_total_qps, row_total_qps);
+        if (max_total_qps < row_total_qps) {
+            max_total_qps = row_total_qps;
+            max_qps_partition_id = row.row_name;
+        }
+
+        // get max_total_cu、min_total_cu and the id of this partition which has max_total_cu
+        double row_total_cu = row.get_total_cu();
+        min_total_cu = std::min(min_total_cu, row_total_cu);
+        if (max_total_cu < row_total_cu) {
+            max_total_cu = row_total_cu;
+            max_cu_partition_id = row.row_name;
+        }
+    }
+
+    void merge(const table_stats &row_stats)
+    {
+        total_get_qps += row_stats.total_get_qps;
+        total_multi_get_qps += row_stats.total_multi_get_qps;
+        total_put_qps += row_stats.total_put_qps;
+        total_multi_put_qps += row_stats.total_multi_put_qps;
+        total_remove_qps += row_stats.total_remove_qps;
+        total_multi_remove_qps += row_stats.total_multi_remove_qps;
+        total_incr_qps += row_stats.total_incr_qps;
+        total_check_and_set_qps += row_stats.total_check_and_set_qps;
+        total_check_and_mutate_qps += row_stats.total_check_and_mutate_qps;
+        total_scan_qps += row_stats.total_scan_qps;
+        total_recent_read_cu += row_stats.total_recent_read_cu;
+        total_recent_write_cu += row_stats.total_recent_write_cu;
+        total_recent_expire_count += row_stats.total_recent_expire_count;
+        total_recent_filter_count += row_stats.total_recent_filter_count;
+        total_recent_abnormal_count += row_stats.total_recent_abnormal_count;
+        total_recent_write_throttling_delay_count +=
+            row_stats.total_recent_write_throttling_delay_count;
+        total_recent_write_throttling_reject_count +=
+            row_stats.total_recent_write_throttling_reject_count;
+        total_storage_mb += row_stats.total_storage_mb;
+        total_storage_count += row_stats.total_storage_count;
+        total_rdb_block_cache_hit_count += row_stats.total_rdb_block_cache_hit_count;
+        total_rdb_block_cache_total_count += row_stats.total_rdb_block_cache_total_count;
+        total_rdb_index_and_filter_blocks_mem_usage +=
+            row_stats.total_rdb_index_and_filter_blocks_mem_usage;
+        total_rdb_memtable_mem_usage += row_stats.total_rdb_memtable_mem_usage;
+
+        // We only need max_total_qps/min_total_qps/max_total_cu/min_total_cu in the same app
+        if (this->app_name == row_stats.app_name) {
+            // get max_total_qps、min_total_qps and id of the partition which has max_total_qps
+            min_total_qps = std::min(min_total_qps, row_stats.min_total_qps);
+            if (max_total_qps < row_stats.max_total_qps) {
+                max_total_qps = row_stats.max_total_qps;
+                max_qps_partition_id = row_stats.max_qps_partition_id;
+            }
+
+            // get max_total_cu、min_total_cu and id of the partition which has max_total_cu
+            min_total_cu = std::min(min_total_cu, row_stats.min_total_cu);
+            if (max_total_cu < row_stats.max_total_cu) {
+                max_total_cu = row_stats.max_total_cu;
+                max_cu_partition_id = row_stats.max_cu_partition_id;
+            }
+        }
+    }
+
+    std::string app_name;
+    double total_get_qps = 0;
+    double total_multi_get_qps = 0;
+    double total_put_qps = 0;
+    double total_multi_put_qps = 0;
+    double total_remove_qps = 0;
+    double total_multi_remove_qps = 0;
+    double total_incr_qps = 0;
+    double total_check_and_set_qps = 0;
+    double total_check_and_mutate_qps = 0;
+    double total_scan_qps = 0;
+    double total_recent_read_cu = 0;
+    double total_recent_write_cu = 0;
+    double total_recent_expire_count = 0;
+    double total_recent_filter_count = 0;
+    double total_recent_abnormal_count = 0;
+    double total_recent_write_throttling_delay_count = 0;
+    double total_recent_write_throttling_reject_count = 0;
+    double total_storage_mb = 0;
+    double total_storage_count = 0;
+    double total_rdb_block_cache_hit_count = 0;
+    double total_rdb_block_cache_total_count = 0;
+    double total_rdb_index_and_filter_blocks_mem_usage = 0;
+    double total_rdb_memtable_mem_usage = 0;
+    double max_total_qps = 0;
+    double min_total_qps = INT_MAX;
+    double max_total_cu = 0;
+    double min_total_cu = INT_MAX;
+    std::string max_qps_partition_id;
+    std::string max_cu_partition_id;
+};
